+++ src/App.vue
@@ -26,4 +26,11 @@ export default {
   margin: 0;
   padding: 0;
 }
+.icon {
+  width: 1em;
+  height: 1em;
+  vertical-align: -0.15em;
+  fill: currentColor;
+  overflow: hidden;
+}
 </style>