+++ src/components/CustomAccountPubComponent/switch.vue
@@ -1,8 +1,8 @@
 <template>
-  <div :style="style">
+  <div :style="style" @click="$emit('click')">
     <div class="switch">
       <div class="num">{{ data.num }}</div>
-      <div class="length">{{ data.length }}</div>
+      <!-- <div class="length">{{ data.length }}</div> -->
       <div class="bar">
         <svg class="icon" aria-hidden="true">
           <use xlink:href="#icon-a-zhexian1"></use>
@@ -14,7 +14,7 @@
 
 <script>
 export default {
-  name: 'switch',
+  name: 'switchComp',
   props: {
     data: {
       type: Object,