+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 --|--|--
  pubComponent page | 增加大数据量图表，波形图页面 | [a8698a1](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/a8698a1)
  pubComponent page | 增加大数据量图表，波形图页面 | [7106c75](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/7106c75)
+ pubComponent pages | 增加注释，增加点击和拖动事件切换 | [b9071bf](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/b9071bf)
  pubComponents pages | 增加播放功能 | [5e73051](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/5e73051)
  pubComponents pages | 增加横向测量，划块测量 | [2267e83](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/2267e83)
+ pubComponents pages | 增加图表类型 | [8257553](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/8257553)
  pubComponents pages | 增加异常查看 | [ec2fb0e](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/ec2fb0e)
  pubComponents | 增加y轴测量功能 | [e3399ae](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/e3399ae)
  publicComponents pages | 增加展示单个通道功能 | [35d1be1](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/35d1be1)