+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.1.0 (2022-08-24)
+# 0.1.0 (2022-08-25)
 
 ### 🌟 新功能
 范围|描述|commitId
@@ -7,6 +7,7 @@
  pubComponent page | 增加大数据量图表，波形图页面 | [a8698a1](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/a8698a1)
  pubComponent page | 增加大数据量图表，波形图页面 | [7106c75](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/7106c75)
  pubComponent pages | 增加注释，增加点击和拖动事件切换 | [b9071bf](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/b9071bf)
+ pubComponents demos | 增加自制台账组件 | [3716728](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/3716728)
  pubComponents pages | 增加播放功能 | [5e73051](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/5e73051)
  pubComponents pages | 增加横向测量，划块测量 | [2267e83](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/2267e83)
  pubComponents pages | 增加图表类型 | [8257553](http://192.168.1.102:9900/yangzhenfeng/tky-front/commits/8257553)